# services/scoring/openapi.yaml
openapi: 3.1.0
info:
  title: OCR Slip Gaji Scoring API
  description: 'Tahap terakhir pipeline slip gaji: memberi **skor keyakinan per field** — P(nilai ini
    benar), 0-1, terkalibrasi isotonic.


    Modelnya regresi logistik 9 variabel + one-hot field, dilatih pada 1.428 nilai dari 50 dokumen berlabel;
    out-of-fold AUC 0,826 · Gini 0,652 · ECE 0,037. Pada ambang 80: precision 93,4%, recall 75,6%.


    **Tidak ada keputusan di sini.** Skor mentah yang dikembalikan; pemetaan ke `confidence` 0/1 terjadi
    di kontrak `extract-ocr`, supaya ambangnya bisa diubah tanpa melatih ulang apa pun.


    **Pipeline asinkron:** service structuring POST /v1/scoring/jobs dan menerima 202; service ini memberi
    skor di latar, menyimpan hasilnya, dan mengirim callback berisi hasil akhir ke orchestrator. /v1/scoring/score
    melakukan pekerjaan yang sama secara sinkron. Semua endpoint kecuali /health memerlukan header X-API-Key.


    ## Conventions (the same for every nilam-ocr service)


    **Authentication.** Every endpoint except `/health`, `/ready` and `/metrics` requires the header `X-API-Key`.

    A missing or wrong key answers `401`. The services are reachable only from inside the cluster.


    **Envelope.** Every JSON response, success or error, has the same shape:

    `{status_code, status_desc, message, data, errors, request_id}`. `status_code` always equals the HTTP

    status. On success `errors` is null; on error `data` is null.
...
servers:
...
paths:
...
  /v1/scoring/score:
    post:
      tags:
      - Scoring
      summary: Skor keyakinan per field, sinkron (tanpa job, tanpa callback)
      description: 'Menghitung **P(nilai field ini benar)** untuk setiap nilai yang diextraction, sebagai
        skor 0-1 yang sudah dikalibrasi (isotonic), lalu dikembalikan apa adanya.


        Modelnya regresi logistik atas 9 variabel + one-hot field: kesepakatan regex vs LLM, jarak nominal
        ke median field, seberapa sering field ini benar-benar ada, jumlah digit, hasil cek aritmetika,
        nilai nol, dan porsi kotak OCR berskor rendah di halaman. Dilatih pada 1.428 nilai dari 50 dokumen
        berlabel; out-of-fold (GroupKFold per dokumen): AUC 0,826 · Gini 0,652 · ECE 0,037.


        **Tidak ada keputusan di sini.** Pemetaan ke `confidence` 0/1 terjadi di kontrak `extract-ocr`
        memakai `threshold`, supaya ambangnya bisa diubah tanpa melatih ulang apa pun. Pada ambang 80:
        precision 93,4%, recall 75,6%.


        Field tanpa nilai tidak diberi skor: tidak ada yang perlu dinilai di sana, dan memberinya 0 akan
        tercampur dengan nilai yang benar-benar diragukan.'
      operationId: scoreSlips
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRequest'
        required: true
      responses:
        '200':
          description: Skor untuk tiap slip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreResponse'
              examples:
                satu_slip:
                  summary: Satu slip
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      slips:
                      - slip_no: 1
                        scores:
                          nama_karyawan: 76.71
                          jabatan: 76.71
                          gaji_pokok: 93.6
                          tunjangan_makan: 88.89
                          total_pendapatan: 92.47
                          gaji_bersih: 92.47
                      threshold: 80.0
                      model: conf_v2
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '400':
          description: Tidak ada slip untuk dinilai
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 400
                status_desc: Bad Request
                message: tidak ada slip untuk dinilai
                errors: tidak ada slip untuk dinilai
                pipeline_last_stage: scoring
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
                pipeline_last_stage: scoring
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'body.slips: Field required'
                errors: VALIDATION_ERROR
                pipeline_last_stage: scoring
        '500':
          description: Model keyakinan gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 500
                status_desc: Internal Server Error
                message: Internal error in SCORING stage
                errors: Internal error in SCORING stage
                pipeline_last_stage: scoring
      security:
      - APIKeyHeader: []
...
components:
  schemas:
    ErrorResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          examples:
          - 400
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          examples:
          - Bad Request
        message:
          type: string
          title: Message
          description: Human-readable cause; safe to log
          examples:
          - Uploaded file is empty
        data:
          type: 'null'
          title: Data
          description: Always null on error
        errors:
          type: string
          title: Errors
          description: Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message
            names a different field every time); for every other status it repeats `message`
          examples:
          - Uploaded file is empty
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Present when the error is tied to a known request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - status_code
      - status_desc
      - message
      - errors
      title: ErrorResponse
      description: Every error response.
...
    ScoreRequest:
      properties:
        slips:
          items:
            $ref: '#/components/schemas/SlipPayload'
          type: array
          minItems: 1
          title: Slips
          description: 'Slip hasil tahap structuring, apa adanya. `source`, `checks`, `counts`, `llm.compare`,
            `ocr_text` dan `ocr_confidence` harus ikut: fitur terkuat model justru berasal dari jejak
            itu — kesepakatan regex vs LLM, jarak nominal ke median field, dan porsi kotak OCR berskor
            rendah di halaman'
      type: object
      required:
      - slips
      title: ScoreRequest
    ScoreResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/ScoringResult'
      type: object
      required:
      - data
      title: ScoreResponse
...
    ScoringResult:
      properties:
        slips:
          items:
            $ref: '#/components/schemas/SlipScores'
          type: array
          title: Slips
          description: Satu entri per slip
        threshold:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Threshold
          description: 'Ambang bawaan (0-1) yang memisahkan `confidence: 1` dari `0` di kontrak `extract-ocr`
            (`FIELD_CONFIDENCE_THRESHOLD`); `column_confidence_threshold` permintaan menimpanya per field.
            Skor mentah tetap dikembalikan, supaya ambangnya bisa diubah tanpa melatih ulang apa pun'
          examples:
          - 0.5
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: Backend model yang memberi skor
          examples:
          - conf_v2
        column_confidence_threshold:
          anyOf:
          - additionalProperties:
              type: number
            type: object
          - type: 'null'
          title: Column Confidence Threshold
          description: Ambang per field dari permintaan (`column_confidence_threshold`); null bila tidak
            dikirim
          examples:
          - all_field: 0.6
            gaji_bersih: 0.9
      type: object
      required:
      - slips
      - threshold
      title: ScoringResult
...
    SlipPayload:
      properties:
        slip_no:
          type: integer
          minimum: 1.0
          title: Slip No
          description: Nomor urut slip dalam dokumen
          examples:
          - 1
        page:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Page
          description: Halaman asal slip ini
          examples:
          - 1
        fields:
          additionalProperties: true
          type: object
          title: Fields
          description: 20 field, selalu lengkap; null bila tidak ditemukan. Nominal uang sudah integer
        source:
          additionalProperties: true
          type: object
          title: Source
          description: 'Asal tiap nilai (`how`: regex | llm | derived, sinonim label, baris OCR) — fitur
            model keyakinan'
        checks:
          additionalProperties: true
          type: object
          title: Checks
          description: Hasil cek aritmetika slip
        counts:
          additionalProperties: true
          type: object
          title: Counts
          description: Berapa field terisi, dan dari lapisan mana
        llm:
          additionalProperties: true
          type: object
          title: Llm
          description: Jejak arbitrase LLM; `compare` dibaca model keyakinan
        ocr_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Ocr Text
          description: Teks halaman asal — fitur mutu OCR di tahap scoring
        ocr_confidence:
          additionalProperties: true
          type: object
          title: Ocr Confidence
          description: Skor OCR halaman asal
        missing_mandatory_fields:
          items:
            type: string
          type: array
          title: Missing Mandatory Fields
          description: Field wajib yang tidak terisi pada slip ini
      additionalProperties: true
      type: object
      required:
      - slip_no
      - fields
      title: SlipPayload
      description: Satu slip gaji hasil structuring.
    SlipScores:
      properties:
        slip_no:
          type: integer
          minimum: 1.0
          title: Slip No
          examples:
          - 1
        scores:
          additionalProperties:
            type: number
          type: object
          title: Scores
          description: '{field: P(nilai ini benar), 0-1}. Hanya field yang ada nilainya yang diberi skor'
          examples:
          - gaji_pokok: 0.936
            total_pendapatan: 0.9247
            tunjangan_makan: 0.8889
      type: object
      required:
      - slip_no
      - scores
      title: SlipScores
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: API key for this service
      in: header
      name: X-API-Key
...